rename osplus to ostools
