+++ exo-go/src/applicationrunner/initializer.go
@@ -10,7 +10,7 @@ import (
 	"github.com/Originate/exosphere/exo-go/src/dockercompose"
 	"github.com/Originate/exosphere/exo-go/src/dockercomposebuilder"
 	"github.com/Originate/exosphere/exo-go/src/logger"
-	"github.com/Originate/exosphere/exo-go/src/osplus"
+	"github.com/Originate/exosphere/exo-go/src/ostools"
 )
 
 // Initializer sets up the app
@@ -91,7 +91,7 @@ func (i *Initializer) RenderDockerCompose(dockerComposeDir string) error {
 	if err != nil {
 		return err
 	}
-	if err := osplus.CreateEmptyDirectory(dockerComposeDir); err != nil {
+	if err := ostools.CreateEmptyDirectory(dockerComposeDir); err != nil {
 		return err
 	}
 	return ioutil.WriteFile(path.Join(dockerComposeDir, "docker-compose.yml"), bytes, 0777)
