+++ exo-go/src/cmd/add.go
@@ -6,7 +6,7 @@ import (
 	"path"
 
 	"github.com/Originate/exosphere/exo-go/src/config"
-	"github.com/Originate/exosphere/exo-go/src/osplus"
+	"github.com/Originate/exosphere/exo-go/src/ostools"
 	"github.com/Originate/exosphere/exo-go/src/template"
 	prompt "github.com/segmentio/go-prompt"
 	"github.com/spf13/cobra"
@@ -41,7 +41,7 @@ var addCmd = &cobra.Command{
 		if err != nil {
 			panic(err)
 		}
-		subdirectories, err := osplus.GetSubdirectories(serviceTmpDir)
+		subdirectories, err := ostools.GetSubdirectories(serviceTmpDir)
 		if err != nil {
 			panic(err)
 		}
@@ -55,11 +55,11 @@ var addCmd = &cobra.Command{
 			fmt.Println(err)
 			os.Exit(1)
 		}
-		err = osplus.MoveDirectory(path.Join(serviceTmpDir, serviceRole), path.Join(appDir, serviceRole))
+		err = ostools.MoveDirectory(path.Join(serviceTmpDir, serviceRole), path.Join(appDir, serviceRole))
 		if err != nil {
 			panic(err)
 		}
-		if !osplus.DoesFileExist(path.Join(appDir, serviceRole, "service.yml")) {
+		if !ostools.DoesFileExist(path.Join(appDir, serviceRole, "service.yml")) {
 			var templateDir string
 			templateDir, err = template.CreateServiceTemplateDir(serviceRole)
 			if err != nil {
