+++ exo-go/src/cmd/deploy.go
@@ -8,7 +8,7 @@ import (
 	"github.com/Originate/exosphere/exo-go/src/applicationdeployer"
 	"github.com/Originate/exosphere/exo-go/src/config"
 	"github.com/Originate/exosphere/exo-go/src/logger"
-	"github.com/Originate/exosphere/exo-go/src/osplus"
+	"github.com/Originate/exosphere/exo-go/src/ostools"
 	"github.com/spf13/cobra"
 )
 
@@ -26,7 +26,7 @@ var deployCmd = &cobra.Command{
 		if err != nil {
 			panic(err)
 		}
-		homeDir, err := osplus.GetHomeDirectory()
+		homeDir, err := ostools.GetHomeDirectory()
 		if err != nil {
 			panic(err)
 		}
