+++ exo-go/src/cmd/run.go
@@ -9,7 +9,7 @@ import (
 	"github.com/Originate/exosphere/exo-go/src/applicationrunner"
 	"github.com/Originate/exosphere/exo-go/src/config"
 	"github.com/Originate/exosphere/exo-go/src/logger"
-	"github.com/Originate/exosphere/exo-go/src/osplus"
+	"github.com/Originate/exosphere/exo-go/src/ostools"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
@@ -26,7 +26,7 @@ var runCmd = &cobra.Command{
 		if err != nil {
 			panic(err)
 		}
-		homeDir, err := osplus.GetHomeDirectory()
+		homeDir, err := ostools.GetHomeDirectory()
 		if err != nil {
 			panic(err)
 		}
