+++ exo-go/src/ostools/does_directory_exist.go
@@ -1,4 +1,4 @@
-package osplus
+package ostools
 
 // DoesDirectoryExist returns true if the directory dirPath is an existing directory,
 // and false otherwise
