+++ exo-go/src/ostools/does_file_exist.go
@@ -1,4 +1,4 @@
-package osplus
+package ostools
 
 import "os"
