+++ exo-go/src/ostools/is_empty_directory.go
@@ -1,4 +1,4 @@
-package osplus
+package ostools
 
 import (
 	"io"
